+++ nvflare/apis/server_engine_spec.py
@@ -16,7 +16,6 @@ from abc import ABC, abstractmethod
 from typing import Dict, List, Optional, Tuple
 
 from nvflare.apis.shareable import Shareable
-from nvflare.apis.streaming import ConsumerFactory, ObjectProducer, StreamContext
 from nvflare.widgets.widget import Widget
 
 from .client import Client
@@ -164,67 +163,6 @@ class ServerEngineSpec(EngineSpec, ABC):
         self, targets: [], topic: str, request: Shareable, fl_ctx: FLContext, optional=False, secure=False
     ) -> dict:
         return self.send_aux_request(targets, topic, request, 0.0, fl_ctx, optional, secure=secure)
-
-    @abstractmethod
-    def stream_objects(
-        self,
-        channel: str,
-        topic: str,
-        stream_ctx: StreamContext,
-        targets: List[str],
-        producer: ObjectProducer,
-        fl_ctx: FLContext,
-        optional=False,
-        secure=False,
-    ):
-        """Send a stream of Shareable objects to receivers.
-
-        Args:
-            channel: the channel for this stream
-            topic: topic of the stream
-            stream_ctx: context of the stream
-            targets: receiving sites
-            producer: the ObjectProducer that can produces the stream of Shareable objects
-            fl_ctx: the FLContext object
-            optional: whether the stream is optional
-            secure: whether to use P2P security
-
-        Returns: result from the generator's reply processing
-
-        """
-        pass
-
-    @abstractmethod
-    def shutdown_streamer(self):
-        pass
-
-    @abstractmethod
-    def register_stream_processing(
-        self,
-        channel: str,
-        topic: str,
-        factory: ConsumerFactory,
-        stream_done_cb=None,
-        **cb_kwargs,
-    ):
-        """Register a ConsumerFactory for specified app channel and topic.
-        Once a new streaming request is received for the channel/topic, the registered factory will be used
-        to create an ObjectConsumer object to handle the new stream.
-
-        Note: the factory should generate a new ObjectConsumer every time get_consumer() is called. This is because
-        multiple streaming sessions could be going on at the same time. Each streaming session should have its
-        own ObjectConsumer.
-
-        Args:
-            channel: app channel
-            topic: app topic
-            factory: the factory to be registered
-            stream_done_cb: the callback to be called when streaming is done on receiving side
-
-        Returns: None
-
-        """
-        pass
 
     @abstractmethod
     def get_widget(self, widget_id: str) -> Widget:
