+++ nvflare/private/fed/client/client_engine_executor_spec.py
@@ -20,7 +20,6 @@ from nvflare.apis.client_engine_spec import ClientEngineSpec
 from nvflare.apis.engine_spec import EngineSpec
 from nvflare.apis.fl_context import FLContext
 from nvflare.apis.shareable import Shareable
-from nvflare.apis.streaming import ConsumerFactory, ObjectProducer, StreamContext
 from nvflare.apis.workspace import Workspace
 from nvflare.widgets.widget import Widget
 
@@ -157,67 +156,6 @@ class ClientEngineExecutorSpec(ClientEngineSpec, EngineSpec, ABC):
             secure: whether to send the message in P2P secure mode
 
         Returns:
-
-        """
-        pass
-
-    @abstractmethod
-    def stream_objects(
-        self,
-        channel: str,
-        topic: str,
-        stream_ctx: StreamContext,
-        targets: List[str],
-        producer: ObjectProducer,
-        fl_ctx: FLContext,
-        optional=False,
-        secure=False,
-    ):
-        """Send a stream of Shareable objects to receivers.
-
-        Args:
-            channel: the channel for this stream
-            topic: topic of the stream
-            stream_ctx: context of the steam
-            targets: receiving sites
-            producer: the ObjectProducer that can produce the stream of Shareable objects
-            fl_ctx: the FLContext object
-            optional: whether the stream is optional
-            secure: whether to use P2P security
-
-        Returns: result from the producer's reply processing
-
-        """
-        pass
-
-    @abstractmethod
-    def shutdown_streamer(self):
-        pass
-
-    @abstractmethod
-    def register_stream_processing(
-        self,
-        channel: str,
-        topic: str,
-        factory: ConsumerFactory,
-        stream_done_cb=None,
-        **cb_kwargs,
-    ):
-        """Register a ConsumerFactory for specified app channel and topic
-        Once a new streaming request is received for the channel/topic, the registered factory will be used
-        to create an ObjectConsumer object to consume objects of the stream.
-
-        Note: the factory should generate a new ObjectConsumer every time get_consumer() is called. This is because
-        multiple streaming sessions could be going on at the same time. Each streaming session should have its
-        own ObjectConsumer.
-
-        Args:
-            channel: app channel
-            topic: app topic
-            factory: the factory to be registered
-            stream_done_cb: callback to be called when streaming is done on receiving side
-
-        Returns: None
 
         """
         pass
