+++ nvflare/private/fed/client/client_run_manager.py
@@ -20,7 +20,7 @@ from nvflare.apis.fl_component import FLComponent
 from nvflare.apis.fl_constant import FLContextKey, ProcessType, ServerCommandKey, ServerCommandNames, SiteType
 from nvflare.apis.fl_context import FLContext, FLContextManager
 from nvflare.apis.shareable import Shareable
-from nvflare.apis.streaming import ConsumerFactory, ObjectProducer, StreamContext
+from nvflare.apis.streaming import ConsumerFactory, ObjectProducer, StreamableEngine, StreamContext
 from nvflare.apis.workspace import Workspace
 from nvflare.fuel.f3.cellnet.core_cell import FQCN
 from nvflare.fuel.f3.cellnet.defs import MessageHeaderKey
@@ -58,7 +58,7 @@ class ClientRunInfo(object):
 GET_CLIENTS_RETRY = 300
 
 
-class ClientRunManager(ClientEngineExecutorSpec):
+class ClientRunManager(ClientEngineExecutorSpec, StreamableEngine):
     """ClientRunManager provides the ClientEngine APIs implementation running in the child process (CJ)."""
 
     def __init__(
@@ -348,6 +348,9 @@ class ClientRunManager(ClientEngineExecutorSpec):
         optional=False,
         secure=False,
     ):
+        if not self.object_streamer:
+            raise RuntimeError("object streamer has not been created")
+
         return self.object_streamer.stream(
             channel=channel,
             topic=topic,
@@ -367,10 +370,14 @@ class ClientRunManager(ClientEngineExecutorSpec):
         stream_done_cb=None,
         **cb_kwargs,
     ):
+        if not self.object_streamer:
+            raise RuntimeError("object streamer has not been created")
+
         self.object_streamer.register_stream_processing(channel, topic, factory, stream_done_cb, **cb_kwargs)
 
     def shutdown_streamer(self):
-        self.object_streamer.shutdown()
+        if self.object_streamer:
+            self.object_streamer.shutdown()
 
     def abort_app(self, job_id: str, fl_ctx: FLContext):
         runner = fl_ctx.get_prop(key=FLContextKey.RUNNER, default=None)
