+++ tests/unit_test/app_common/job_schedulers/job_scheduler_test.py
@@ -24,7 +24,6 @@ from nvflare.apis.job_def_manager_spec import JobDefManagerSpec
 from nvflare.apis.job_scheduler_spec import DispatchInfo
 from nvflare.apis.resource_manager_spec import ResourceManagerSpec
 from nvflare.apis.server_engine_spec import ServerEngineSpec
-from nvflare.apis.streaming import ConsumerFactory, ObjectProducer, StreamContext
 from nvflare.app_common.job_schedulers.job_scheduler import DefaultJobScheduler
 from nvflare.app_common.resource_managers.list_resource_manager import ListResourceManager
 
@@ -124,32 +123,6 @@ class MockServerEngine(ServerEngineSpec):
         optional: bool = False,
         secure: bool = False,
     ) -> dict:
-        pass
-
-    def stream_objects(
-        self,
-        channel: str,
-        topic: str,
-        stream_ctx: StreamContext,
-        targets: List[str],
-        producer: ObjectProducer,
-        fl_ctx: FLContext,
-        optional=False,
-        secure=False,
-    ):
-        pass
-
-    def register_stream_processing(
-        self,
-        channel: str,
-        topic: str,
-        factory: ConsumerFactory,
-        stream_done_cb=None,
-        **cb_kwargs,
-    ):
-        pass
-
-    def shutdown_streamer(self):
         pass
 
     def get_widget(self, widget_id: str):
